Refactor EditableImage
Replace redundant code to display an image editable by the user with a component that does the same thing
